Split tests/test_monitors.py into per-monitor files.
